Make g_earthquakeTime into a uint16_t, remove DukePlayer_t's name[] member.
Bump BYTEVERSION.

git-svn-id: https://svn.eduke32.com/eduke32@3478 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/gameexec.c
@@ -3112,7 +3112,7 @@ nullquote:
 
         case CON_QUAKE:
             insptr++;
-            g_earthquakeTime = (uint8_t)Gv_GetVarX(*insptr++);
+            g_earthquakeTime = Gv_GetVarX(*insptr++);
             A_PlaySound(EARTHQUAKE,g_player[screenpeek].ps->i);
             continue;
